deplacement du fichier de ryan, ajout du fichier yannis, commentaire Readme
