+++ README.md
@@ -1,1 +1,8 @@
 # testUnitaire
+
+Ce dossier consiste a creer une application simpliste avec un formulaire et une page de connection.
+
+Ensuite nous devrons creer des tests unitaires pour cahcune des fonctions dans les fichiers.
+
+Merci a Ryan pour la base symfony
+Merci a Axel pour l'aide de mise en place du projet.
